--- model-empty-style.json
@@ -1,267 +1,0 @@
-{
-  "version":8,
-  "name":"Model Style",
-  "metadata":{
-    "mapbox:autocomposite":true
-  },
-  "sources":{
-    "composite":{
-      "url":"mapbox://mapbox.mapbox-terrain-v2,mapbox.mapbox-streets-v7",
-      "type":"vector"
-    }
-  },
-  "sprite":"mapbox://sprites/mapbox/streets-v8",
-  "glyphs":"mapbox://fonts/mapbox/{fontstack}/{range}.pbf",
-  "layers":[
-    {
-      "filter":[
-        "==",
-        "class",
-        "national_park"
-      ],
-      "id":"national_park",
-      "source-layer":"landuse_overlay"
-    },
-    {
-      "filter":[
-        "==",
-        "class",
-        "park"
-      ],
-      "id":"parks",
-      "source-layer":"landuse"
-    },
-    {
-      "filter":[
-        "==",
-        "structure",
-        "tunnel"
-      ],
-      "type":"line",
-
-      "id":"tunnel",
-      "source-layer":"road"
-    },
-    {
-      "filter":[
-        "!in",
-        "structure",
-        "bridge",
-        "tunnel"
-      ],
-      "type":"line",
-
-      "id":"roads",
-      "source-layer":"road"
-    },
-    {
-      "filter":[
-        "==",
-        "structure",
-        "bridge"
-      ],
-      "type":"line",
-
-      "id":"bridges",
-      "source-layer":"road"
-    },
-    {
-      "filter":[
-        "all",
-        [
-          "in",
-          "class",
-          "major_rail",
-          "minor_rail"
-        ],
-        [
-          "!in",
-          "structure",
-          "bridge",
-          "tunnel"
-        ]
-      ],
-      "type":"line",
-
-      "id":"road-rail-tracks",
-      "source-layer":"road"
-    },
-    {
-      "filter":[
-        "all",
-        [
-          "==",
-          "maritime",
-          0
-        ],
-        [
-          "==",
-          "disputed",
-          0
-        ]
-      ],
-      "type":"line",
-
-      "id":"admin boundaries",
-      "source-layer":"admin"
-    },
-    {
-      "minzoom":1,
-      "filter":[
-        "all",
-        [
-          "==",
-          "maritime",
-          0
-        ],
-        [
-          "==",
-          "disputed",
-          1
-        ]
-      ],
-      "type":"line",
-
-      "id":"admin dispute boundaries",
-      "source-layer":"admin"
-    },
-    {
-      "filter":[
-        "all",
-        [
-          "!in",
-          "shield",
-          "us-interstate",
-          "us-interstate-business",
-          "us-interstate-duplex",
-          "se-main",
-          "at-expressway",
-          "at-motorway",
-          "at-state-b",
-          "de-motorway",
-          "ch-motorway",
-          "ch-main",
-          "e-road",
-          "hu-main",
-          "hu-motorway",
-          "gr-motorway",
-          "gr-national",
-          "bg-motorway",
-          "bg-national",
-          "hr-motorway",
-          "hr-state",
-          "cz-motorway",
-          "cz-expressway",
-          "cz-road",
-          "fi-main",
-          "nz-state",
-          "pl-motorway",
-          "pl-expressway",
-          "pl-national",
-          "ro-motorway",
-          "ro-national",
-          "ro-county",
-          "sk-highway",
-          "sk-road",
-          "si-motorway",
-          "si-expressway",
-          "za-metropolitan",
-          "za-national",
-          "za-provincial",
-          "za-regional",
-          "rs-motorway",
-          "rs-state-1b"
-        ],
-        [
-          "<=",
-          "reflen",
-          6
-        ]
-      ],
-      "type":"symbol",
-
-      "id":"road-shields-black",
-      "source-layer":"road_label"
-    },
-    {
-      "filter":[
-        "all",
-        [
-          "in",
-          "shield",
-          "us-interstate",
-          "us-interstate-business",
-          "us-interstate-duplex",
-          "se-main",
-          "at-expressway",
-          "at-motorway",
-          "at-state-b",
-          "de-motorway",
-          "ch-motorway",
-          "ch-main",
-          "e-road",
-          "hu-main",
-          "hu-motorway",
-          "gr-motorway",
-          "gr-national",
-          "bg-motorway",
-          "bg-national",
-          "hr-motorway",
-          "hr-state",
-          "cz-motorway",
-          "cz-expressway",
-          "cz-road",
-          "fi-main",
-          "nz-state",
-          "pl-motorway",
-          "pl-expressway",
-          "pl-national",
-          "ro-motorway",
-          "ro-national",
-          "ro-county",
-          "sk-highway",
-          "sk-road",
-          "si-motorway",
-          "si-expressway",
-          "za-metropolitan",
-          "za-national",
-          "za-provincial",
-          "za-regional",
-          "rs-motorway",
-          "rs-state-1b"
-        ],
-        [
-          "<=",
-          "reflen",
-          6
-        ]
-      ],
-      "type":"symbol",
-
-      "id":"road-shields-white",
-      "source-layer":"road_label"
-    },
-    {
-      "filter":[
-        "==",
-        "type",
-        "town"
-      ],
-      "type":"symbol",
-
-      "id":"place town",
-      "source-layer":"place_label"
-    },
-    {
-      "filter":[
-        "==",
-        "type",
-        "city"
-      ],
-      "type":"symbol",
-
-      "id":"place city",
-      "source-layer":"place_label"
-    }
-  ]
-}
