--- test/test_check-layers.js
@@ -1,72 +1,0 @@
-var _ = require('underscore-node');
-var test = require('tape');
-var model = require('../model-streets-style.json');
-// pass in json files
-var style = require('../styles/streets-v8.json');
-
-// the function that does the work
-function styled(hasLayer, checkLayer) {
-  // loop thru every layer in the style!
-
-  var match;
-  var check;
-
-  // first check if we have the 1st layer we are checking for
-  style.layers.forEach(function(layer) {
-    if(layer.id === hasLayer) {
-      // if we have a match, save the value
-      match = model.layers.filter(function(model_layer) {
-        return model_layer.id === layer.id;
-      })[0];
-      // does this have a ref layer?
-      if(match && match.ref) {
-        match = model.layers.filter(function(model_layer) {
-          return model_layer.key === layer.id;
-        })[0];
-      }
-      // now verify the layers have matching filter values
-      if(match) {
-        var keys = Object.keys(match);
-        keys.forEach(function(key) {
-          //t.same(match[key], layer[key], 'This layer passes.');
-        });
-        console.log('The style has ' + layer.id);
-      }
-    }
-  });
-
-  // now check we have the 2nd layer we are looking for
-  style.layers.forEach(function(layer) {
-    if(layer.id === checkLayer) {
-      // if we have a match, save the value
-      check = model.layers.filter(function(model_layer) {
-        return model_layer.id === layer.id;
-      })[0];
-      // does this have a ref layer?
-      if(check && check.ref) {
-        check = model.layers.filter(function(model_layer) {
-          return model_layer.key === layer.id;
-        })[0];
-      }
-      // now verify the layers have matching filter values
-      if(check) {
-        var keys = Object.keys(check);
-        keys.forEach(function(key) {
-          //t.same(check[key], layer[key], 'This layer passes.');
-        });
-        console.log('This style also has ' + layer.id);
-      }
-    }
-  });
-
-  // if there is no match, then tell me
-  if(match === undefined) {
-    console.log('There is no layer called ' + hasLayer);
-  }
-  if(check === undefined) {
-    console.log('There is no layer called ' + checkLayer);
-  }
-}
-
-// run function to check for layers
-styled('parks', 'national_park');
